users/views/login.py
""" User login view """

#Django REST framework
from rest_framework import status
from rest_framework import APIView

#Serializers
from users.serializers.login import UserLoginSerializer


class UserLoginAPIView(APIView):
    pass